+++ Source/Endpoint.swift
@@ -40,28 +40,49 @@ public class Endpoint<Target> {
 
     /// Convenience method for creating a new Endpoint with the same properties as the receiver, but with added parameters.
     public func endpointByAddingParameters(parameters: [String: AnyObject]) -> Endpoint<Target> {
-        var newParameters = self.parameters ?? [String: AnyObject]()
-        for (key, value) in parameters {
-            newParameters[key] = value
-        }
-
-        return Endpoint(URL: URL, sampleResponseClosure: sampleResponseClosure, method: method, parameters: newParameters, parameterEncoding: parameterEncoding, httpHeaderFields: httpHeaderFields)
+        return endpointByAdding(parameters: parameters)
     }
 
     /// Convenience method for creating a new Endpoint with the same properties as the receiver, but with added HTTP header fields.
     public func endpointByAddingHTTPHeaderFields(httpHeaderFields: [String: String]) -> Endpoint<Target> {
-        var newHTTPHeaderFields = self.httpHeaderFields ?? [String: String]()
-        for (key, value) in httpHeaderFields {
-            newHTTPHeaderFields[key] = value
-        }
-
-        return Endpoint(URL: URL, sampleResponseClosure: sampleResponseClosure, method: method, parameters: parameters, parameterEncoding: parameterEncoding, httpHeaderFields: newHTTPHeaderFields)
+        return endpointByAdding(httpHeaderFields: httpHeaderFields)
     }
     
     /// Convenience method for creating a new Endpoint with the same properties as the receiver, but with another parameter encoding.
     public func endpointByAddingParameterEncoding(newParameterEncoding: Moya.ParameterEncoding) -> Endpoint<Target> {
+        return endpointByAdding(parameterEncoding: newParameterEncoding)
+    }
+    
+    /// Convenience method for creating a new Endpoint, with changes only to the properties we specify as parameters
+    public func endpointByAdding(parameters parameters: [String: AnyObject]? = nil, httpHeaderFields: [String: String]? = nil, parameterEncoding: Moya.ParameterEncoding? = nil)  -> Endpoint<Target> {
+        let newParameters = addParameters(parameters)
+        let newHTTPHeaderFields = addHTTPHeaderFields(httpHeaderFields)
+        let newParameterEncoding = parameterEncoding ?? self.parameterEncoding
+        return Endpoint(URL: URL, sampleResponseClosure: sampleResponseClosure, method: method, parameters: newParameters, parameterEncoding: newParameterEncoding, httpHeaderFields: newHTTPHeaderFields)
+    }
+    
+    private func addParameters(parameters: [String: AnyObject]?) -> [String: AnyObject]? {
+        guard let unwrappedParameters = parameters where unwrappedParameters.count > 0 else {
+            return self.parameters
+        }
         
-        return Endpoint(URL: URL, sampleResponseClosure: sampleResponseClosure, method: method, parameters: parameters, parameterEncoding: newParameterEncoding, httpHeaderFields: httpHeaderFields)
+        var newParameters = self.parameters ?? [String: AnyObject]()
+        unwrappedParameters.forEach { (key, value) in
+            newParameters[key] = value
+        }
+        return newParameters
+    }
+    
+    private func addHTTPHeaderFields(headers: [String: String]?) -> [String: String]? {
+        guard let unwrappedHeaders = headers where unwrappedHeaders.count > 0 else {
+            return self.httpHeaderFields
+        }
+        
+        var newHTTPHeaderFields = self.httpHeaderFields ?? [String: String]()
+        unwrappedHeaders.forEach { (key, value) in
+            newHTTPHeaderFields[key] = value
+        }
+        return newHTTPHeaderFields
     }
 }
 
@@ -72,7 +93,7 @@ extension Endpoint {
         request.HTTPMethod = method.rawValue
         request.allHTTPHeaderFields = httpHeaderFields
 
-        return parameterEncoding.toAlamofire.encode(request, parameters: parameters).0
+        return parameterEncoding.encode(request, parameters: parameters).0
     }
 }