+++ Source/Moya+Alamofire.swift
@@ -4,25 +4,7 @@ import Alamofire
 public typealias Manager = Alamofire.Manager
 
 /// Choice of parameter encoding.
-public enum ParameterEncoding {
-    case URL
-    case JSON
-    case PropertyList(NSPropertyListFormat, NSPropertyListWriteOptions)
-    case Custom((URLRequestConvertible, [String: AnyObject]?) -> (NSMutableURLRequest, NSError?))
-    
-    internal var toAlamofire: Alamofire.ParameterEncoding {
-        switch self {
-        case .URL:
-            return .URL
-        case .JSON:
-            return .JSON
-        case .PropertyList(let format, let options):
-            return .PropertyList(format, options)
-        case .Custom(let closure):
-            return .Custom(closure)
-        }
-    }
-}
+public typealias ParameterEncoding = Alamofire.ParameterEncoding
 
 /// Make the Alamofire Request type conform to our type, to prevent leaking Alamofire to plugins.
 extension Request: RequestType { }